Make app description multilingual.
- Force app description to be a dict of the type that can be passed to
  multilingual funcitons for hicp components.

Not a perfect solution, assumes apps know what language codes need to be
supported, and since they don't need to know anything about each other, this
isn't the case.

Better would be able to return a text ID, but each app defines its own IDs
(some dynamically), there is no way to know what ID to use.

So current strategy is to return something that can be used with an arbitrary
ID, or to obtain an ID for it from the text group functions, and count on the
app using it properly.

Possible alternative is, instead of App object returning an AppInfo object
(get_app_info()), add a method (get_app_desc()) which takes hicp object,
updates it with all supported languages, and returns a text ID.

Probably needs more thinking.